My changes

+++ database/migrations/2024_12_05_105342_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->nullable();
             $table->text('description')->nullable();
-            $table->decimal('price')->nullable();
+            $table->double('price')->nullable();
             $table->double('quantity')->nullable();
             $table->text('image')->nullable();
             $table->unsignedBigInteger('store_id');
